% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NestedFuncs.R
\name{fun1}
\alias{fun1}
\title{Test function level 1}
\usage{
fun1(num.sim = 10, num.per = 8, num.day = 5, ...)
}
\arguments{
\item{num.sim}{first variable for function 1}

\item{num.per}{second variable for function 1}

\item{num.day}{third variable for function 1}
}
\description{
Test function level 1
}
