% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NestedFuncs.R
\name{fun3}
\alias{fun3}
\title{Test function level 3}
\usage{
fun3()
}
\arguments{
\item{var31}{first variable for function 3}

\item{fun3on}{turn the formula on or off}
}
\description{
Test function level 3
}
