% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NestedFuncs.R
\name{fun4}
\alias{fun4}
\title{Test function level 4}
\usage{
fun4()
}
\arguments{
\item{var41}{first variable for function 4}

\item{fun4on}{turn the formula on or off}
}
\description{
Test function level 4
}
