% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForeachParamsPass.R
\name{localregargs}
\alias{localregargs}
\title{local regular arguments localregargs function}
\usage{
localregargs(fun2vars, fun3vars, fun4vars)
}
\description{
local regular arguments localregargs function
}
